ws code cleanup
